feat(types): set up proper type exporting & configure other packages to use it
