+++ tsconfig.json
@@ -19,6 +19,7 @@
       "@pokt-foundation/pocketjs": ["packages/pocket"],
       "@pokt-foundation/pocketjs-keybase": ["packages/keybase"],
       "@pokt-foundation/pocketjs-provider": ["packages/provider"],
+      "@pokt-foundation/pocketjs-types": ["packages/types"],
       "@pokt-foundation/pocketjs-utils": ["packages/utils"]
     }
   }
